Implemented notification toggle in settings
