+++ wsdot/AppDelegate.swift
@@ -32,7 +32,7 @@ import Realm
 
 
 @UIApplicationMain
-class AppDelegate: UIResponder, UIApplicationDelegate {
+class AppDelegate: UIResponder, UIApplicationDelegate{
     
     var window: UIWindow?
     let gcmMessageIDKey = "gcm.message_id"
@@ -42,26 +42,12 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
         migrateRealm()
         CachesStore.initCacheItem()
         
-        // Register for remote notifications. This shows a permission dialog on first run, to
-        // show the dialog at a more appropriate time move this registration accordingly.
-        // [START register_for_notifications]
         if #available(iOS 10.0, *) {
-            let authOptions: UNAuthorizationOptions = [.alert, .badge, .sound]
-            UNUserNotificationCenter.current().requestAuthorization(
-                options: authOptions,
-                completionHandler: {_, _ in })
-            
             // For iOS 10 display notification (sent via APNS)
             UNUserNotificationCenter.current().delegate = self
             // For iOS 10 data message (sent via FCM)
             FIRMessaging.messaging().remoteMessageDelegate = self
-        } else {
-            let settings: UIUserNotificationSettings =
-                UIUserNotificationSettings(types: [.alert, .badge, .sound], categories: nil)
-            application.registerUserNotificationSettings(settings)
         }
-        
-        application.registerForRemoteNotifications()
         
         GMSServices.provideAPIKey(ApiKeys.google_key)
         FIRApp.configure()
